+++ handshake/services/DBService/models/config_base.py
@@ -8,12 +8,12 @@ from handshake.services.DBService.models.result_base import RunBase
 from tortoise.models import Model
 from tortoise.fields import (
     CharEnumField,
-    ForeignKeyField,
     ForeignKeyRelation,
     TextField,
     BooleanField,
     JSONField,
     DatetimeField,
+    OneToOneField,
 )
 
 
@@ -28,8 +28,8 @@ class ConfigBase(Model):
 
 
 class TestConfigBase(Model):
-    test: ForeignKeyRelation[RunBase] = ForeignKeyField(
-        "models.RunBase", related_name="runs", to_field="testID"
+    test: ForeignKeyRelation[RunBase] = OneToOneField(
+        "models.RunBase", related_name="runs", to_field="testID", pk=True
     )
     platform = TextField(null=False, description="could be windows or linux")
     framework = TextField(null=False, description="name of the framework used")